Fix an oversight in LG.Term.isSubtermOf (!)
Previously, the function would not recognise subterms when a name of
a (co)variable appeared only in a command or the divisor of a cut.
The remedy was to also compare those names.

Unfortunately, this introduces a subtle semantic bug where the name
of a covariable may match the name of a cut divisor that represents a
variable or vice versa. The only way to fix that problem would be to
use separate name types for values and contexts, which would explode
the Cut constructor from LG.Term.ContextTerm into 6 different
versions.